+++ ChangeLog.md
@@ -4,6 +4,10 @@ ___
 
 # NOT RELEASED
 
+
+
+## RELEASE 3.2 - 29/06/2021
+
 - FIX : Bad fix for pgsql ... *29/06/2021* - 3.2.6
 - FIX : module family *24/06/2021* - 3.2.5
 - NEW : Compatibility V14 - script which aim at replace __REFCLIENT__ with __REF_CLIENT__ *09/06/2021* - 3.2.4